Add automated tests and improve dashboard refresh
