+++ jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  testMatch: ["**/tests/**/*.test.js"],
+  testEnvironment: "node",
+  clearMocks: true
+};
