+++ server/index.js
@@ -1,460 +1,31 @@
-const express = require("express");
-const path = require("path");
-const cors = require("cors");
-const { randomUUID } = require("crypto");
+const { createApp } = require("./app");
+const { initializeDatabase } = require("./initializeDatabase");
 
-const {
-  initializeDatabase,
-  getQuizConfig,
-  runQuery,
-  allQuery,
-  getQuery
-} = require("./initializeDatabase");
-
-const app = express();
 const PORT = process.env.PORT || 3000;
-const SESSION_TIMEOUT_MS = parseInt(process.env.SESSION_TIMEOUT_MS || "60000", 10);
 
-app.use(cors());
-app.use(express.json());
+async function startServer(port = PORT, appOptions = {}) {
+  await initializeDatabase();
+  const app = createApp(appOptions);
 
-app.use(express.static(path.join(__dirname, "..", "public")));
-app.use("/styles", express.static(path.join(__dirname, "..", "styles")));
-app.use("/src", express.static(path.join(__dirname, "..", "src")));
-
-function asyncHandler(fn) {
-  return (req, res, next) => {
-    Promise.resolve(fn(req, res, next)).catch(next);
-  };
-}
-
-function sanitizeOptions(options = []) {
-  if (!Array.isArray(options)) {
-    return [];
-  }
-  return options
-    .map((option) => ({
-      id: option.id || randomUUID(),
-      label: option.label || "",
-      isCorrect: Boolean(option.isCorrect || option.is_correct)
-    }))
-    .filter((option) => option.label.trim().length > 0);
-}
-
-async function getModuleById(moduleId) {
-  return getQuery("SELECT * FROM modules WHERE id = ?", [moduleId]);
-}
-
-async function getQuestionWithOptions(questionId) {
-  const question = await getQuery(
-    "SELECT * FROM questions WHERE id = ?",
-    [questionId]
-  );
-  if (!question) {
-    return null;
-  }
-
-  const options = await allQuery(
-    "SELECT * FROM options WHERE question_id = ? ORDER BY position ASC",
-    [questionId]
-  );
-
-  return {
-    ...question,
-    options: options.map((option) => ({
-      id: option.id,
-      label: option.label,
-      isCorrect: Boolean(option.is_correct)
-    }))
-  };
-}
-
-async function saveQuestionOptions(questionId, options) {
-  await runQuery("DELETE FROM options WHERE question_id = ?", [questionId]);
-
-  for (let index = 0; index < options.length; index += 1) {
-    const option = options[index];
-    await runQuery(
-      `INSERT INTO options (id, question_id, label, is_correct, position)
-       VALUES (?, ?, ?, ?, ?)`,
-      [
-        option.id || randomUUID(),
-        questionId,
-        option.label || "",
-        option.isCorrect ? 1 : 0,
-        index
-      ]
-    );
-  }
-}
-
-app.get(
-  "/api/quiz-config",
-  asyncHandler(async (req, res) => {
-    const config = await getQuizConfig();
-    res.json(config);
-  })
-);
-
-app.get(
-  "/api/modules",
-  asyncHandler(async (req, res) => {
-    const modules = await allQuery(
-      "SELECT id, title, questions_per_session AS questionsPerSession FROM modules ORDER BY position ASC"
-    );
-    res.json(modules);
-  })
-);
-
-app.get(
-  "/api/questions",
-  asyncHandler(async (req, res) => {
-    const rows = await allQuery(
-      `SELECT q.id, q.text, q.type, q.module_id AS moduleId, q.position,
-              m.title AS moduleTitle
-       FROM questions q
-       INNER JOIN modules m ON q.module_id = m.id
-       ORDER BY m.position ASC, q.position ASC`
-    );
-    res.json(rows);
-  })
-);
-
-app.get(
-  "/api/questions/:id",
-  asyncHandler(async (req, res) => {
-    const question = await getQuestionWithOptions(req.params.id);
-    if (!question) {
-      res.status(404).json({ message: "Vraag niet gevonden" });
-      return;
-    }
-    res.json({
-      id: question.id,
-      moduleId: question.module_id,
-      text: question.text,
-      type: question.type,
-      feedback: {
-        correct: question.feedback_correct || "",
-        incorrect: question.feedback_incorrect || ""
-      },
-      options: question.options
-    });
-  })
-);
-
-app.post(
-  "/api/questions",
-  asyncHandler(async (req, res) => {
-    const { moduleId, text, type = "single", feedback = {}, options = [] } =
-      req.body || {};
-
-    if (!moduleId || !text) {
-      res.status(400).json({ message: "moduleId en text zijn verplicht" });
-      return;
-    }
-
-    const module = await getModuleById(moduleId);
-    if (!module) {
-      res.status(404).json({ message: "Module niet gevonden" });
-      return;
-    }
-
-    const sanitizedOptions = sanitizeOptions(options);
-    if (!sanitizedOptions.length) {
-      res.status(400).json({ message: "Voeg minimaal één antwoordoptie toe" });
-      return;
-    }
-
-    const hasCorrect = sanitizedOptions.some((option) => option.isCorrect);
-    if (!hasCorrect) {
-      res.status(400).json({ message: "Markeer minimaal één juist antwoord" });
-      return;
-    }
-
-    const now = Date.now();
-    const positionRow = await getQuery(
-      "SELECT COALESCE(MAX(position), -1) AS maxPosition FROM questions WHERE module_id = ?",
-      [moduleId]
-    );
-    const nextPosition = (positionRow?.maxPosition || -1) + 1;
-
-    const questionId = randomUUID();
-    await runQuery(
-      `INSERT INTO questions (id, module_id, text, type, feedback_correct, feedback_incorrect, position)
-       VALUES (?, ?, ?, ?, ?, ?, ?)`,
-      [
-        questionId,
-        moduleId,
-        text,
-        type,
-        feedback.correct || "",
-        feedback.incorrect || "",
-        nextPosition
-      ]
-    );
-
-    await saveQuestionOptions(questionId, sanitizedOptions);
-
-    res.status(201).json({ id: questionId });
-  })
-);
-
-app.put(
-  "/api/questions/:id",
-  asyncHandler(async (req, res) => {
-    const { moduleId, text, type = "single", feedback = {}, options = [] } =
-      req.body || {};
-
-    if (!moduleId || !text) {
-      res.status(400).json({ message: "moduleId en text zijn verplicht" });
-      return;
-    }
-
-    const existing = await getQuestionWithOptions(req.params.id);
-    if (!existing) {
-      res.status(404).json({ message: "Vraag niet gevonden" });
-      return;
-    }
-
-    const module = await getModuleById(moduleId);
-    if (!module) {
-      res.status(404).json({ message: "Module niet gevonden" });
-      return;
-    }
-
-    const sanitizedOptions = sanitizeOptions(options);
-    if (!sanitizedOptions.length) {
-      res.status(400).json({ message: "Voeg minimaal één antwoordoptie toe" });
-      return;
-    }
-
-    const hasCorrect = sanitizedOptions.some((option) => option.isCorrect);
-    if (!hasCorrect) {
-      res.status(400).json({ message: "Markeer minimaal één juist antwoord" });
-      return;
-    }
-
-    await runQuery(
-      `UPDATE questions
-       SET module_id = ?, text = ?, type = ?, feedback_correct = ?, feedback_incorrect = ?
-       WHERE id = ?`,
-      [
-        moduleId,
-        text,
-        type,
-        feedback.correct || "",
-        feedback.incorrect || "",
-        req.params.id
-      ]
-    );
-
-    await saveQuestionOptions(req.params.id, sanitizedOptions);
-
-    res.json({ id: req.params.id });
-  })
-);
-
-app.post(
-  "/api/sessions",
-  asyncHandler(async (req, res) => {
-    const name = req.body?.name || null;
-    const requestedId = req.body?.id;
-    const now = new Date();
-    const id = requestedId || randomUUID();
-    const iso = now.toISOString();
-
-    await runQuery(
-      `INSERT INTO sessions (id, name, status, start_time, last_seen)
-       VALUES (?, ?, 'active', ?, ?)`,
-      [id, name, iso, iso]
-    );
-
-    res.status(201).json({
-      id,
-      name,
-      status: "active",
-      startTime: iso,
-      lastSeen: iso
-    });
-  })
-);
-
-app.post(
-  "/api/sessions/:id/heartbeat",
-  asyncHandler(async (req, res) => {
-    const id = req.params.id;
-    const now = new Date().toISOString();
-    await runQuery(
-      `UPDATE sessions
-       SET last_seen = ?, status = CASE WHEN status = 'completed' THEN status ELSE 'active' END
-       WHERE id = ?`,
-      [now, id]
-    );
-    res.json({ lastSeen: now });
-  })
-);
-
-app.post(
-  "/api/sessions/:id/attempt",
-  asyncHandler(async (req, res) => {
-    const id = req.params.id;
-    const { moduleId, questionId, selectedOptionIds = [], isCorrect = false } =
-      req.body || {};
-
-    const now = new Date().toISOString();
-    await runQuery(
-      `INSERT INTO session_attempts (session_id, module_id, question_id, selected_options, is_correct, answered_at)
-       VALUES (?, ?, ?, ?, ?, ?)`,
-      [
-        id,
-        moduleId || null,
-        questionId || null,
-        JSON.stringify(selectedOptionIds || []),
-        isCorrect ? 1 : 0,
-        now
-      ]
-    );
-
-    await runQuery(
-      `UPDATE sessions
-       SET last_seen = ?, status = CASE WHEN status = 'completed' THEN status ELSE 'active' END
-       WHERE id = ?`,
-      [now, id]
-    );
-
-    res.status(201).json({});
-  })
-);
-
-app.post(
-  "/api/sessions/:id/complete",
-  asyncHandler(async (req, res) => {
-    const id = req.params.id;
-    const summary = req.body?.summary || null;
-    const now = new Date().toISOString();
-
-    await runQuery(
-      `UPDATE sessions
-       SET status = 'completed', summary = ?, end_time = ?, last_seen = ?
-       WHERE id = ?`,
-      [summary ? JSON.stringify(summary) : null, now, now, id]
-    );
-
-    res.json({ endTime: now });
-  })
-);
-
-app.post(
-  "/api/sessions/:id/leave",
-  asyncHandler(async (req, res) => {
-    const id = req.params.id;
-    const now = new Date().toISOString();
-    await runQuery(
-      `UPDATE sessions
-       SET status = CASE WHEN status = 'completed' THEN status ELSE 'inactive' END,
-           end_time = COALESCE(end_time, ?),
-           last_seen = ?
-       WHERE id = ?`,
-      [now, now, id]
-    );
-    res.json({});
-  })
-);
-
-app.get(
-  "/api/dashboard",
-  asyncHandler(async (req, res) => {
-    const startOfDay = new Date();
-    startOfDay.setHours(0, 0, 0, 0);
-    const startIso = startOfDay.toISOString();
-
-    const sessions = await allQuery(
-      `SELECT * FROM sessions WHERE start_time >= ? ORDER BY start_time DESC`,
-      [startIso]
-    );
-
-    if (!sessions.length) {
-      res.json({
-        totalSessions: 0,
-        totalCorrect: 0,
-        totalIncorrect: 0,
-        activeParticipants: 0,
-        activeSessions: []
-      });
-      return;
-    }
-
-    const sessionIds = sessions.map((session) => session.id);
-    const attemptRows = await allQuery(
-      `SELECT session_id AS sessionId,
-              SUM(CASE WHEN is_correct = 1 THEN 1 ELSE 0 END) AS correct,
-              SUM(CASE WHEN is_correct = 0 THEN 1 ELSE 0 END) AS incorrect
-       FROM session_attempts
-       WHERE session_id IN (${sessionIds.map(() => "?").join(",")})
-       GROUP BY session_id`,
-      sessionIds
-    );
-
-    const attemptsMap = attemptRows.reduce((acc, row) => {
-      acc[row.sessionId] = {
-        correct: row.correct || 0,
-        incorrect: row.incorrect || 0
-      };
-      return acc;
-    }, {});
-
-    const cutoff = new Date(Date.now() - SESSION_TIMEOUT_MS).toISOString();
-
-    let totalCorrect = 0;
-    let totalIncorrect = 0;
-    const activeSessions = [];
-
-    sessions.forEach((session) => {
-      const stats = attemptsMap[session.id] || { correct: 0, incorrect: 0 };
-      totalCorrect += stats.correct;
-      totalIncorrect += stats.incorrect;
-
-      if (
-        session.status === "active" &&
-        session.last_seen &&
-        session.last_seen >= cutoff
-      ) {
-        activeSessions.push({
-          id: session.id,
-          name: session.name,
-          correct: stats.correct,
-          incorrect: stats.incorrect,
-          startTime: session.start_time,
-          lastSeen: session.last_seen
-        });
-      }
-    });
-
-    res.json({
-      totalSessions: sessions.length,
-      totalCorrect,
-      totalIncorrect,
-      activeParticipants: activeSessions.length,
-      activeSessions
-    });
-  })
-);
-
-app.use((err, req, res, next) => {
-  // eslint-disable-next-line no-console
-  console.error(err);
-  res.status(500).json({ message: "Er is iets misgegaan" });
-});
-
-initializeDatabase()
-  .then(() => {
-    app.listen(PORT, () => {
+  return new Promise((resolve, reject) => {
+    const server = app.listen(port, () => {
       // eslint-disable-next-line no-console
-      console.log(`Server gestart op poort ${PORT}`);
+      console.log(`Server gestart op poort ${port}`);
+      resolve({ app, server });
     });
-  })
-  .catch((error) => {
+
+    server.on("error", (error) => {
+      reject(error);
+    });
+  });
+}
+
+if (require.main === module) {
+  startServer().catch((error) => {
     // eslint-disable-next-line no-console
-    console.error("Database-initialisatie mislukt", error);
+    console.error("Database-initialisatie of serverstart mislukt", error);
     process.exit(1);
   });
+}
+
+module.exports = { createApp, startServer };
